refactor(api)!: content mutations move to /api/content
A blog post is an Article row with Type == "blog", so create, replace, delete and
every workflow transition are the same operation either way. They all lived under
/api/articles, which meant editing a blog post was POST /api/articles/{id}/edit
and the e2e sweeper deleted blog rows through /articles/{id}. The route said
"article" while the handler meant "content".

    POST   /content                          was  POST   /articles
    PUT    /content/{id}                     was  PUT    /articles/{id}
    DELETE /content/{id}                     was  DELETE /articles/{id}
    POST   /content/{id}/publish             ...and the other five transitions

Reads stay split by type — /articles, /articles/{slug}, /review/articles against
their /blog counterparts — because a reader genuinely is asking for one or the
other.

The handlers move file too, into ContentFunctions. Filing them under
ArticlesFunctions was the same untruth one level down. ArticlesFunctions keeps
the three reads and its constructor narrows to (IContentRepository repo) — no
sanitizer, no logger — which leaves it structurally identical to BlogFunctions.
That symmetry is the point of the split.

VisibleInReview moves to ArticleVisibility rather than travelling with the
mutations: BlogFunctions calls it, and a pure-read class must not depend on the
mutations class.

No aliases. Every caller is in-repo and moves here; unlike the subscribe rename
these are all role-gated, so an unknown external caller is unlikely.

The route inventory added earlier did its job — the move showed up as nine
changed rows plus two hard-coded EntryPoint literals in JwtMiddlewareTests, which
name a namespace-qualified method path as a string and so would otherwise have
failed with the gate silently resolving to null.

Two e2e call sites would have failed silently rather than loudly, and are the
reason for the hand-grep: content-revision's waitForResponse matcher (a stale
predicate becomes a 30-second hang, not an assertion failure) and seed.ts's
sweeper, whose del() does not check its status and would have leaked every
fixture into subsequent runs. permissions.spec now asserts positively that the
old routes 404, so "the route vanished" cannot pass as "nobody called it".

error-surfacing.spec needs no changes: all three of its interceptions target read
routes that stay, and its fourth matcher is an endsWith('/publish') suffix.

Verified: build -warnaserror clean, 392 API tests; web lint clean, 466 unit
tests, vue-tsc clean. E2E not run locally — the harness cannot start Azurite on
this machine.
