+++ src/api/Slypn.Api.Tests/RouteInventoryTests.cs
@@ -62,17 +62,17 @@ public class RouteInventoryTests
     // Sorted by route then method so the diff reads like the Swagger page.
     private static readonly string[] Expected =
     [
-        "DELETE /articles/{id}                            DeleteArticle              Admin                        Slypn.Api.Functions.ArticlesFunctions.Delete",
+        "DELETE /content/{id}                            DeleteArticle              Admin                        Slypn.Api.Functions.ContentFunctions.Delete",
         "GET    /articles                                 GetArticles                optional-auth                Slypn.Api.Functions.ArticlesFunctions.GetArticles",
-        "POST   /articles                                 CreateArticle              Admin+Contributor            Slypn.Api.Functions.ArticlesFunctions.Create",
+        "POST   /content                                 CreateArticle              Admin+Contributor            Slypn.Api.Functions.ContentFunctions.Create",
         "GET    /articles/{slug}                          GetArticleBySlug           optional-auth                Slypn.Api.Functions.ArticlesFunctions.GetArticleBySlug",
-        "PUT    /articles/{id}                            ReplaceArticle             Admin+Contributor            Slypn.Api.Functions.ArticlesFunctions.Replace",
-        "POST   /articles/{id}/cancel-deletion            CancelArticleDeletion      Admin                        Slypn.Api.Functions.ArticlesFunctions.CancelDeletion",
-        "POST   /articles/{id}/edit                       EditPublishedArticle       Admin+Contributor            Slypn.Api.Functions.ArticlesFunctions.Edit",
-        "POST   /articles/{id}/publish                    PublishArticle             Admin                        Slypn.Api.Functions.ArticlesFunctions.Publish",
-        "POST   /articles/{id}/request-deletion           RequestArticleDeletion     Admin+Contributor            Slypn.Api.Functions.ArticlesFunctions.RequestDeletion",
-        "POST   /articles/{id}/revise                     ReviseArticle              Admin                        Slypn.Api.Functions.ArticlesFunctions.Revise",
-        "POST   /articles/{id}/withdraw                   WithdrawArticle            Admin+Contributor            Slypn.Api.Functions.ArticlesFunctions.Withdraw",
+        "PUT    /content/{id}                            ReplaceArticle             Admin+Contributor            Slypn.Api.Functions.ContentFunctions.Replace",
+        "POST   /content/{id}/cancel-deletion            CancelArticleDeletion      Admin                        Slypn.Api.Functions.ContentFunctions.CancelDeletion",
+        "POST   /content/{id}/edit                       EditPublishedArticle       Admin+Contributor            Slypn.Api.Functions.ContentFunctions.Edit",
+        "POST   /content/{id}/publish                    PublishArticle             Admin                        Slypn.Api.Functions.ContentFunctions.Publish",
+        "POST   /content/{id}/request-deletion           RequestArticleDeletion     Admin+Contributor            Slypn.Api.Functions.ContentFunctions.RequestDeletion",
+        "POST   /content/{id}/revise                     ReviseArticle              Admin                        Slypn.Api.Functions.ContentFunctions.Revise",
+        "POST   /content/{id}/withdraw                   WithdrawArticle            Admin+Contributor            Slypn.Api.Functions.ContentFunctions.Withdraw",
         "GET    /blog                                     GetBlogPosts               optional-auth                Slypn.Api.Functions.BlogFunctions.GetBlogPosts",
         "GET    /blog/{slug}                              GetBlogPostBySlug          optional-auth                Slypn.Api.Functions.BlogFunctions.GetBlogPostBySlug",
         "GET    /review/articles                          GetPendingArticles         Admin+Contributor            Slypn.Api.Functions.ArticlesFunctions.GetPendingArticles",
